pyproject.toml
[tool.poetry]
name = "tubecaster-telegram-bot"
version = "0.0.16"
description = "Simple video to audio caster telegram bot"
authors = ["lilliputten <lilliputten@gmail.com>"]
readme = "README.md"

# @see https://python-poetry.org/docs/pyproject/#packages
packages = [
    { include = "api" },
]

[tool.poetry.dependencies]
# NOTE: ffmpeg is required, see: https://www.ffmpeg.org/download.html
python = "^3.11"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pyTelegramBotAPI = "^4.24.0" # The telebot package to control telegram bots
yt-dlp = "^2024.11.18" # The video/audio dowload library
concurrent-log-handler = "^0.9.25" # Write logs to a reotating files
# NOTE: Using external ffmpeg for audio processing, see `core/ffmpeg`
# librosa = "^0.10.2.post1" # NOTE: ffmpeg is required, see: https://www.ffmpeg.org/download.html
# pydub = "^0.25.1"
# ffmpeg-python = "^0.2.0" # @see https://kkroening.github.io/ffmpeg-python/
prisma = "^0.15.0"
colorama = "^0.4.6"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
# Don't use newer versions if pyright: they bring a lot of false typings' errors
pyright = "==1.1.403"
poetry-plugin-export = "^1.8.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 120
# ignore = "E501,W6"  # or ["E501", "W6"]
in-place = true
recursive = true
aggressive = 3

[tool.black]
line-length = 120

[tool.blue]
line-length = 120
# TODO!

[tool.poetry.scripts]
export_requirements = "poetry-scripts:export_requirements"
lint = "poetry-scripts:lint"
format = "poetry-scripts:format"
check_all = "poetry-scripts:check_all"
# test = "poetry-scripts:test"
prisma_format = "poetry-scripts.prisma:format"
prisma_format_and_validate = "poetry-scripts.prisma:format_and_validate"
prisma_db_push = "poetry-scripts.prisma:db_push"
prisma_db_push_test = "poetry-scripts.prisma:db_push_test"
prisma_db_push_all = "poetry-scripts.prisma:db_push_all"
